requirements.txt
astropy>=5.0
numpy>=1.20
treecorr>=5.0
pyyaml>=6.0
matplotlib>=3.5
